Make the local build truly local
We kept running into problems with using the root of the source
directory as the home directory for the simulated build. Emacs kept
putting turds all over the root directory in the form of `save-list`,
`elpa`, `quelpa`, etc... Addittionally, there was no way to use the
quelpa fetcher to set the package source directory as the repo root
directory. It just wouldn't do it.

The answer then, is to just make a `/build` directory that can house
all of the sources and all of the temp files, and can be swept away
simply. So now, the only thing that is ignored is `build`

```
build/home -> contains the simulated home directory
build/lib -> all of the frontmacs elisp sources
```

If the `FRONTMACS_RUNLOCAL` variable is defined, then the provided
`init.el` will try and use the soures from the file system, otherwise
it will use the latest on github.