formalize boot process.
It was unclear how you would actually use frontmacs in your
day-to-day. We have a package, but how do you actually use it?

This init script sets up all the prerequisites needed to install
frontmacs, installs it if necessary, and then invokes it. The idea is
that you just need to copy this script into your `.emacs.d` directory,
and then load it from your `init.el`.

```
  (load "init-frontmacs.el")
```

With that single line, you're off to the races.

As a bonus, the init script recognizes a FRONTMACS_ARCHIVE environment
variable to use a local archive useful for local development.