Add configuration and initialization mechanism
Frontmacs is curated, but it is also needs to be customizable. This
allows users to scratch that itch of uniqueness even inside the walled
garden of a great out-of-the-box experience. It's also important to have
a customization mechanism in place so that we can begin using Frontmacs,
but still have a clear way to fill in the gaps that are going to be
large at first.

To enable this customization, this is introduces the concept of
configuration and initializers. The configuration file (config.el) is
for setting variables that will affect the Frontmacs initialization
process, and it will be generated for you the first time you install
it.

The other capability added is for initializers. Any `.el` file that is
dropped into `.emacs.d/initializers` will be evaluated after frontmacs
has loaded, and so it can be used as an all-purpose setup
mechanism... an escape hatch for anything that Frontmacs hasn't covered
yet. In this way, it's just like the normal `init.el` mechanism except
it allows you to logically divide up your initializers into separate
files according to purpose. E.g.

`.emacs.d/initializers/ruby.el`
`.emacs.d/initializers/javascript.el`
`.emacs.d/initializers/github.el`