Explicity map in, out, and unzoo functions.
While the zoom and unzoom functions seem to work, the "unzoom"
function never remembers exactly where it is when using the mac
keys. This probably has something to do with the strange
`(control ?0)` form which I've never seen before.

In either case, this explicit mapping is every bit as clear, if not
clearer, and unzooming to the default zoom level works as expected.