Pin js2-mode to MELPA
js2-mode is now available on both the MELPA Stable and MELPA, and so
it's always choosing melpa stable. However, many of the packages on
which we depend, depend on later versions of js2-mode, and so an
install from scratch will no longer work.

This sets js2-mode to always load from MELPA no matter what.