Add support for typescript mode.
TypeScript is amazing, and TIDE has all the nice refactorations that
can be used within TS mode.

I included it in JavaScript setup because it's all JS under the
covers, and also because perhaps eventually we'll switch to using TIDE
for editting JavaScript code as well.

Currently there are no bindings other than `M-.` for Jump to Symbol,
and funcitonality has to be referenced by name.