Use vanilla `rspec` executable to run specs.
There was a period where everybody seemed to want to run their rspec
suite from within their `Rakefile`, but the setup and teardown
facilities of rspec itself ought to be sufficient to run your entire
suite. Otherwise, what the hell are you doing anyway?

This makes the default to run rspec with rspec, like rspec intended.