Tell cask explicitly where emacs is living.
There's a bug in python getenv that prepends `/usr/bin` to the path,
which was futzing up cask's logic for which emacs to use. Previously,
there was no `/usr/bin/emacs` so it wasn't a problem. Now however,
there is, and so when python appended `/usr/bin` to PATH it overrode
our version of emacs under test living at $HOME/bin

This explicitly sets the CASK_EMACS environment variable so that cask
will look there and nowhere else.