Better line matching in a node stack trace.
Node has kinda wacky stack traces that can be heterogenous in
formatting. For example:

```
    at /Users/cowboyd/Code/microstates.js/play.mjs:54:14
    at Object.<anonymous> (/Users/cowboyd/Code/microstates.js/play.mjs)
    at Module._compile (module.js:624:30)
```

Notice how the first line has the file:line:colum unwrapped, and the
third has it wrapped in `()`. While the second line doesn't have a
line number or column!

This improves the error regexp to handle both the first and third
line, while still punting on the third.