Disable js2-mode errors
By default, js2-mode uses it's own style linter to display warnings
and errors, but these days almost every repo ships with a linting
solution. It's far safer to assume that there will be a lint config
than that the errors we see will be useful.

Let's just disable the parse errors and warnings (note this will not
effect the sytax highlighting)