Update README to provide more Why than HOW.
The README does a decent job of explaining HOW, but doesn't do a good
job of explainin WHY. WHY is definitely there in the mix, but it isn't
the primary motivator.

To address this, I changed the wording to focus around
individuals (not the Frontside "we") and the problems they face which
are solved by Frontmacs.