Add support for language server.
The [Language Server Protocol][1] is one of the most exciting
developments for language tools in the last two decades. It enables
amazing code insights and change management tools for any IDE.

This introduces basic support for LSP in emacs. At the moment, in
order to turn it on for a buffer, you need to manually run `M-x
lsp`. However, as we experiment with it, it's very likely that we'lll
end up using it automatically for modes.

[1]: https://en.wikipedia.org/wiki/Language_Server_Protocol