Support for EditorConfig
From the EditorConfig website:

> EditorConfig helps maintain consistent coding styles for multiple
> developers working on the same project across various editors and
> IDEs. The EditorConfig project consists of a file format for
> defining coding styles and a collection of text editor plugins that
> enable editors to read the file format and adhere to defined
> styles. EditorConfig files are easily readable and they work nicely
> with version control systems.

When used by a project, it's a great way to invert control on
formatting so that you don't run into stylistic issues at the linting
phase. Especially for things other than source code files like JSON,
YAML, or .xml

This adds support for editor config and turns it on globally. It does
not disply this minor mode on the modeline.