Make it more friendly with use-package
It's redundant to have to declare the use package and then also
require it.

```elisp
(use-package frontside-javascript :config (frontside-javascript))
```

Instead, we'd like the fact that you're using `use-package` explicitly
to declare your intent to actually use the package. In otherwords, be
able to say:

```elisp
(use-package frontside-javascrip)
```

This defines, via autoload, a hook that will be run automatically by
`use-package` when frontside-javascript loads. This saves us the
typing.

Note that the [`use-package-inject-hooks`][1] variable must be
non-nil.

[1]: https://github.com/jwiegley/use-package/blob/master/use-package-core.el#L257-L281