Set JavasScript indent to two spaces
The convention almost everywhere is to use 2 spaces for indent in
JavaScript, however the default in js-mode and js2-mode is 4. All
things being equal, we should follow that convention.

This sets the default to 2 spaces, but still allows it to be
overridden. However, most JavaScript projects, if they do not have
this default, set their own defaults