Support Deno
Deno is on the horizon, and so we want to be able to transition
seamlessly between different types of javascript projects.

Deno ships with its own language server, and so we want to use
TIDE. That means that the choices are emacs-lsp, or eglot. I've opted
for LSP since it's more visual studio like, but we can change to eglot
later.

If we detect that this is a Deno project, then we load lsp-mode for
all javascript file types, rather than js2 or tide.